Implement full CELT decoder with PVQ, IMDCT, and de-emphasis
Replace the simplified CELT stub with a complete RFC 6716 Section 4.3
implementation:

- celt_tables.go: All spec constants — eBands, Laplace model params,
  band allocation table, MDCT window, de-emphasis coefficient
- celt_energy.go: Laplace-coded coarse energy decode, fine energy
  decode with priority finalization
- celt_alloc.go: Bit allocation engine distributing bits to bands
- celt_pvq.go: Pyramid Vector Quantization decoder — U(n,k)/V(n,k)
  tables, cwrsi codeword-to-vector, normalize, exp_rotation
- celt_mdct.go: IMDCT (type-IV DCT), overlap-add with 120-point
  window, de-emphasis IIR filter
- celt.go: Full decode orchestration wiring all stages together
- celt_test.go: Tests for PVQ enumeration, cwrsi, normalization,
  IMDCT, de-emphasis, window monotonicity, Laplace, silence decode

Also adds OUTPUT_FORMAT watershed (ogg vs wav) to cmd/disrecord and
fixes voice connection to handle nil AEAD for test paths.

+++ pkg/opus/decoder_test.go
@@ -73,7 +73,7 @@ func TestDecoder_PLCFade(t *testing.T) {
 	}
 	// Successive PLC calls should produce fading output.
 	_, _ = d.Decode(SilenceFrame, 960, false)
-	d.lastGoodSamples = make([]float64, 960)
+	d.lastGoodSamples = make([]float32, 960)
 	for i := range d.lastGoodSamples {
 		d.lastGoodSamples[i] = 0.5
 	}
